requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.8.0
google-cloud-pubsub==2.21.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
pillow==10.2.0
python-dotenv==1.0.0
websockets==12.0

